Version bump and get clippy's help

See merge request terrence_too/openfa!156

+++ libs/asset/mmm/src/lib.rs
@@ -644,10 +644,7 @@ impl MValue {
 
         for tmap in tmaps.iter() {
             if let TLoc::Index(i) = tmap.1.loc {
-                ensure!(
-                    (i as usize) < tdics.len(),
-                    "expected at tdict for each tmap index"
-                );
+                ensure!(i < tdics.len(), "expected at tdict for each tmap index");
             }
         }
